requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
httpx==0.26.0
apscheduler==3.10.4
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
pydantic-settings==2.1.0
